Autre écriture de setSalaire dans Technicien.java

+++ src/main/java/com/ipiecoles/java/java220/Technicien.java
@@ -29,6 +29,7 @@ public class Technicien extends Employe{
 
     @Override
     public void setSalaire(Double salaire) {
-        super.setSalaire(salaire * (1+(double) grade / 10));
+        //super.setSalaire(salaire * (1 + (double) grade / 10));
+        super.setSalaire(salaire * (1 + grade / 10d));
     }
 }
